Move filters to their own package
