+++ neuro_san/internals/graph/filters/dictionary_common_defs_config_filter.py
@@ -14,7 +14,7 @@ from typing import Dict
 
 from copy import deepcopy
 
-from neuro_san.internals.graph.registry.abstract_common_defs_config_filter \
+from neuro_san.internals.graph.filters.abstract_common_defs_config_filter \
     import AbstractCommonDefsConfigFilter
